+++ src/sdk/sdk.ts
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { PeprValidateRequest } from "./lib/validate-request";
-import { PeprMutateRequest } from "./lib/mutate-request";
-import { a } from "./lib";
+import { PeprValidateRequest } from "../lib/validate-request";
+import { PeprMutateRequest } from "../lib/mutate-request";
+import { a } from "../lib";
 import { V1OwnerReference } from "@kubernetes/client-node";
 import { GenericKind } from "kubernetes-fluent-client";
 import { K8s, kind } from "kubernetes-fluent-client";
-import Log from "./lib/logger";
+import Log from "../lib/logger";
 
 /**
  * Returns all containers in a pod
